Update kinematics, add full body simulation
Allow kinematics to be more flexible to input type. Remove need to
call invert on matrics and instead calculate manually.

Add full simulation with all 6 legs and body. This accounts for
changing the height and footprint radius. The simulation does not yet
handle rotations of the main body relative to the footprint.

Add animations and an animation manager to transition between
animations. This is very rudimentary for now and does not cleanly
transition between animations. Future versions will need to track and
maintain feet that are touching the ground.